+++ source/blender/nodes/function/nodes/node_fn_axes_to_rotation.cc
@@ -146,8 +146,10 @@ class AxesToRotationFunction : public mf::MultiFunction {
 
   uint64_t hash() const override
   {
-    return get_default_hash(
-        9875347984, primary_axis_.as_int(), secondary_axis_.as_int(), tertiary_axis_.as_int());
+    return get_default_hash(int64_t(9875347984),
+                            primary_axis_.as_int(),
+                            secondary_axis_.as_int(),
+                            tertiary_axis_.as_int());
   }
 };
